backend/python/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
psycopg2-binary
asyncpg
sqlalchemy
pgvector
sentence-transformers
rank-bm25
numpy
torch
torchvision
pillow
requests
jieba
beautifulsoup4
lxml
pybloom-live
redis[hiredis]>=4.5.0

